Load profile from database not API
The API attepts to create a profile if it can't find one, which involves a call to workflow. If workflow then tries to load the profile from the API then this causes an infinite recursive loop of the two APIs calling one another. That is bad.

Instead load the user's profile from the database directly on workflow requests.