Use post-status hook for autoresolved tasks
Where a task autoresolves, but the resolver fails then it needs to stay resolved be cause there's no way for a user to retry the task. This means subsequent attempts to make the same change are blocked by the open (but invisible) task.

Switch from a pre-status hook to a post-status for autoresolved so the task will not stay open if resolution fails.