Use patch for updating models
Allows a more explicit setting of the fields we're interested in without risking overwriting other data points.

There's a couple of places in the `update-data` hooks where it goes a level deeper that I wasn't quite prepared to tamper with right now so those have been left as they were.