Replace whereNotIn queries with a whereIn
The `whereNotIn` clause in a query can't hit the database indexes on the status column and so the queries run far slower than other task-list queries.

Replace the `whereNotIn` with a `whereIn` used in conjunction with a filter against a list of excluded statuses so that the indexes are used for the query.