Suppress "resubmitted" option in project next steps
If a project grant or update task is back with a user then force them down the journey of "edit and resubmit" rather than allowing direct resubmission.

The actual available next steps need to be as they are because when it eventually submits it submits with a status of "resubmitted" and so the validator needs to allow that.