Whitelist progress states in task list builder
If you request a task list with a broken progress parameter in the query string then it returns the default underlying query which returns _everything_ with no status filters (including autoresolved) which then breaks in the client.

Instead return an empty list immediately if the progress parameter doesn't make an expected value.