Expand profile tasks endpoint
Adds pagination to the endpoint and allows non-open tasks to be included in the results if a query string parameter is passed.

Also includes tasks opened by the user as well as about the user in the response. This is acceptable in terms of the impact on its other use in preventing profile mergers as we probably shouldn't allow mergers if there are open tasks created by a user anyway because it will mess with "return to applicant" functionality.