Don't attempt to re-index objects with undefined id
The search re-index hook is set up to fire on task submission to ASRU (to handle draft PPL submissions) however entity creation tasks don't have an id at this point and result in an error in the search indexer.

Ignore anything that doesn't have an id and don't try to index it.