Don't auto-resolve reopening a ROP
Set to `resolved` rather then `autoresolved` so that a record of the reopening is visible in the task history for the project.

Also allow the client to be a lot tidier by mapping the statuses of `resolved` ROPs tasks in a decorator rather than doing it in lots of places in the client.

Finally, include the ROP year in the task `modelData`.